Extract shared party-board avatar helper
